extract BundleLoader
