+++ src/debugger/browser/types/BrowserFlavor.ts
@@ -3,6 +3,6 @@
  */
 
 /**
- * PowerPlatform PCF Debugger for VS Code will try to open the Microsoft Edge flavors in the following order: Stable, Beta, Dev and Canary.
+ * The PCF debugger will try to open the Microsoft Edge flavors in the following order: Stable, Beta, Dev and Canary.
  */
 export type BrowserFlavor = "Default" | "Stable" | "Beta" | "Dev" | "Canary";
