+++ src/debugger/browser/types/IBrowserPath.ts
