+++ src/debugger/browser/types/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ */
+
+export { BrowserFlavor } from "./BrowserFlavor";
+export { Platform } from "./Platform";
+export { IBrowserPath } from "./IBrowserPath";
