+++ src/debugger/extension.ts
@@ -2,10 +2,6 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  */
 
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- */
-
 import * as vscode from "vscode";
 
 import { LaunchDebugProvider } from "./configuration";
